adding NewTektonConfig to tests
