+++ deploy/olm-catalog/codeready-toolchain-operator/0.0.1/codeready-toolchain-operator.v0.0.1.clusterserviceversion.yaml
@@ -6,6 +6,24 @@ metadata:
       [
         {
           "apiVersion": "toolchain.openshift.dev/v1alpha1",
+          "kind": "TektonInstallation",
+          "metadata": {
+            "name": "tekton-installation"
+          },
+          "spec": {},
+          "status": {
+            "conditions": [
+              {
+                "lastTransitionTime": "2019-12-16T09:20:19Z",
+                "reason": "Installed",
+                "status": "True",
+                "type": "TektonReady"
+              }
+            ]
+          }
+        },
+        {
+          "apiVersion": "toolchain.openshift.dev/v1alpha1",
           "kind": "CheInstallation",
           "metadata": {
             "name": "toolchain-cheinstallation"
@@ -22,24 +40,6 @@ metadata:
                 "reason": "Installed",
                 "status": "True",
                 "type": "CheReady"
-              }
-            ]
-          }
-        },
-        {
-          "apiVersion": "toolchain.openshift.dev/v1alpha1",
-          "kind": "TektonInstallation",
-          "metadata": {
-            "name": "tekton-installation"
-          },
-          "spec": {},
-          "status": {
-            "conditions": [
-              {
-                "lastTransitionTime": "2019-12-16T09:20:19Z",
-                "reason": "Installed",
-                "status": "True",
-                "type": "TektonReady"
               }
             ]
           }
