+++ hack/deploy_csv.yaml
@@ -215,6 +215,24 @@ data:
             [
               {
                 "apiVersion": "toolchain.openshift.dev/v1alpha1",
+                "kind": "TektonInstallation",
+                "metadata": {
+                  "name": "tekton-installation"
+                },
+                "spec": {},
+                "status": {
+                  "conditions": [
+                    {
+                      "lastTransitionTime": "2019-12-16T09:20:19Z",
+                      "reason": "Installed",
+                      "status": "True",
+                      "type": "TektonReady"
+                    }
+                  ]
+                }
+              },
+              {
+                "apiVersion": "toolchain.openshift.dev/v1alpha1",
                 "kind": "CheInstallation",
                 "metadata": {
                   "name": "toolchain-cheinstallation"
@@ -231,24 +249,6 @@ data:
                       "reason": "Installed",
                       "status": "True",
                       "type": "CheReady"
-                    }
-                  ]
-                }
-              },
-              {
-                "apiVersion": "toolchain.openshift.dev/v1alpha1",
-                "kind": "TektonInstallation",
-                "metadata": {
-                  "name": "tekton-installation"
-                },
-                "spec": {},
-                "status": {
-                  "conditions": [
-                    {
-                      "lastTransitionTime": "2019-12-16T09:20:19Z",
-                      "reason": "Installed",
-                      "status": "True",
-                      "type": "TektonReady"
                     }
                   ]
                 }
